# packages/service-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Starfiniti FURS API",
    "version": "0.1.0",
    "description": "Unofficial self-hosted Slovenian FURS fiscalization service."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/v1/fiscal-invoices": {
      "post": {
        "operationId": "createFiscalInvoice",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FiscalInvoiceRequest"
              },
              "example": {
                "legalEntityId": "019ff57a-f30d-7290-9bb9-951bed760201",
                "taxNumber": "12345678",
                "kind": "STANDARD",
                "businessPremiseId": "TRGOVINA1",
                "electronicDeviceId": "BLAG1",
                "issueLocalTime": "2026-08-12T12:34:56",
                "messageId": "4e64a93a-40fa-4c02-afb1-488534b85e4c",
                "numberingStructure": "B",
                "invoiceAmount": "12.20",
                "paymentAmount": "12.20",
                "taxesPerSeller": [
                  {
                    "vat": [
                      {
                        "rate": "22.00",
                        "taxableAmount": "10.00",
                        "taxAmount": "2.20"
                      }
                    ]
                  }
                ],
                "operator": {
                  "kind": "slovenian",
                  "taxNumber": "87654321"
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Durably accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiscalDocument"
                },
                "example": {
                  "id": "a10dd547-4497-4384-9ed8-3ce39b0ee5d6",
                  "operationClass": "FISCAL_INVOICE",
                  "kind": "STANDARD",
                  "status": "CONFIRMED",
                  "businessPremiseId": "TRGOVINA1",
                  "electronicDeviceId": "BLAG1",
                  "invoiceSequence": "42",
                  "issueLocalTime": "2026-08-12T12:34:56",
                  "messageId": "4e64a93a-40fa-4c02-afb1-488534b85e4c",
                  "payloadSha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                  "schemaSha256": "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
                  "zoi": "1da0adb4cc87fd85f909e0acb99a2aa4",
                  "eor": "b5f1c310-3dea-4331-82f8-d2dc72d9b018",
                  "confirmedAt": "2026-08-12T10:34:57.000Z",
                  "createdAt": "2026-08-12T10:34:56.000Z",
                  "updatedAt": "2026-08-12T10:34:57.000Z",
                  "subsequentSubmit": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Fiscalize permission required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Idempotency or state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Request rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal service failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/fiscal-invoices/{id}": {
      "get": {
        "operationId": "getFiscalInvoice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fiscal document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiscalDocument"
                },
                "example": {
                  "id": "a10dd547-4497-4384-9ed8-3ce39b0ee5d6",
                  "operationClass": "FISCAL_INVOICE",
                  "kind": "STANDARD",
                  "status": "CONFIRMED",
                  "businessPremiseId": "TRGOVINA1",
                  "electronicDeviceId": "BLAG1",
                  "invoiceSequence": "42",
                  "issueLocalTime": "2026-08-12T12:34:56",
                  "messageId": "4e64a93a-40fa-4c02-afb1-488534b85e4c",
                  "payloadSha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                  "schemaSha256": "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
                  "zoi": "1da0adb4cc87fd85f909e0acb99a2aa4",
                  "eor": "b5f1c310-3dea-4331-82f8-d2dc72d9b018",
                  "confirmedAt": "2026-08-12T10:34:57.000Z",
                  "createdAt": "2026-08-12T10:34:56.000Z",
                  "updatedAt": "2026-08-12T10:34:57.000Z",
                  "subsequentSubmit": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Fiscalize permission required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Idempotency or state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Request rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal service failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FURS_SCHEMA_VALIDATION",
                    "message": "Request does not satisfy the fiscal service contract",
                    "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Use a read credential for GET/HEAD routes or a distinct fiscalize/write credential for all mutating routes."
      }
    },
    "schemas": {
...
      "FiscalDocument": {
        "$id": "FiscalDocument",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "operationClass",
          "kind",
          "status",
          "businessPremiseId",
          "issueLocalTime",
          "messageId",
          "payloadSha256",
          "schemaSha256",
          "createdAt",
          "updatedAt",
          "subsequentSubmit"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$"
          },
          "operationClass": {
            "type": "string",
            "enum": [
              "FISCAL_INVOICE",
              "BUSINESS_PREMISE"
            ]
          },
          "kind": {
            "type": "string",
            "enum": [
              "STANDARD",
              "CORRECTION",
              "CANCELLATION",
              "PREMISE"
            ]
          },
          "correctionOfDocumentId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "READY",
              "SENDING",
              "CONFIRMED",
              "ISSUED_WITHOUT_EOR",
              "RETRY_PENDING",
              "REJECTED",
              "MANUAL_REVIEW",
              "REVERSED"
            ]
          },
          "businessPremiseId": {
            "type": "string"
          },
          "electronicDeviceId": {
            "type": "string"
          },
          "invoiceSequence": {
            "type": "string",
            "pattern": "^[1-9]\\d*$"
          },
          "issueLocalTime": {
            "type": "string"
          },
          "messageId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$"
          },
          "payloadSha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "schemaSha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "zoi": {
            "type": "string",
            "pattern": "^[a-f0-9]{32}$"
          },
          "eor": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$"
          },
          "confirmedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "subsequentSubmit": {
            "type": "boolean"
          }
        },
        "examples": [
          {
            "id": "a10dd547-4497-4384-9ed8-3ce39b0ee5d6",
            "operationClass": "FISCAL_INVOICE",
            "kind": "STANDARD",
            "status": "CONFIRMED",
            "businessPremiseId": "TRGOVINA1",
            "electronicDeviceId": "BLAG1",
            "invoiceSequence": "42",
            "issueLocalTime": "2026-08-12T12:34:56",
            "messageId": "4e64a93a-40fa-4c02-afb1-488534b85e4c",
            "payloadSha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
            "schemaSha256": "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
            "zoi": "1da0adb4cc87fd85f909e0acb99a2aa4",
            "eor": "b5f1c310-3dea-4331-82f8-d2dc72d9b018",
            "confirmedAt": "2026-08-12T10:34:57.000Z",
            "createdAt": "2026-08-12T10:34:56.000Z",
            "updatedAt": "2026-08-12T10:34:57.000Z",
            "subsequentSubmit": false
          }
        ]
      },
      "ErrorResponse": {
        "$id": "ErrorResponse",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "correlationId"
            ],
            "properties": {
              "code": {
                "type": "string",
                "pattern": "^[A-Z0-9_:-]{1,100}$"
              },
              "message": {
                "type": "string",
                "maxLength": 500
              },
              "correlationId": {
                "type": "string",
                "pattern": "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$"
              }
            }
          }
        },
        "examples": [
          {
            "error": {
              "code": "FURS_SCHEMA_VALIDATION",
              "message": "Request does not satisfy the fiscal service contract",
              "correlationId": "8b1d74bc-d132-46a0-81fa-b0c07bb8e557"
            }
          }
        ]
      },
...
    }
  }
}